+++ datumaro/cli/contexts/source.py
@@ -10,7 +10,9 @@ from datumaro.components.environment import Environment
 from datumaro.components.errors import ProjectNotFoundError
 from datumaro.util.scope import on_error_do, scope_add, scoped
 
-from ..util import MultilineFormatter, add_subparser, join_cli_args
+from ..util import (
+    MultilineFormatter, add_subparser, join_cli_args, show_video_import_warning,
+)
 from ..util.errors import CliException
 from ..util.project import generate_next_name, load_project
 
@@ -102,7 +104,7 @@ def import_command(args):
     try:
         project = scope_add(load_project(args.project_dir))
     except ProjectNotFoundError:
-        if not show_plugin_help and args.project_dir:
+        if not show_plugin_help:
             raise
 
     if project is not None:
@@ -129,6 +131,9 @@ def import_command(args):
         name = generate_next_name(
             list(project.working_tree.sources) + os.listdir(),
             'source', sep='-', default='1')
+
+    if fmt == 'video_frames':
+        show_video_import_warning()
 
     project.import_source(name, url=args.url, format=args.format,
         options=extra_args, no_cache=True, no_hash=True, rpath=args.rpath)
@@ -229,7 +234,7 @@ def add_command(args):
     try:
         project = scope_add(load_project(args.project_dir))
     except ProjectNotFoundError:
-        if not show_plugin_help and args.project_dir:
+        if not show_plugin_help:
             raise
 
     if project is not None:
@@ -247,6 +252,9 @@ def add_command(args):
             " by providing an Extractor and Importer plugins" % fmt)
 
     extra_args = arg_parser.parse_cmdline(args.extra_args)
+
+    if fmt == 'video_frames':
+        show_video_import_warning()
 
     name, _ = project.add_source(args.path,
         format=args.format, options=extra_args, rpath=args.rpath)