+++ datumaro/cli/util/__init__.py
@@ -4,6 +4,7 @@
 
 from typing import Iterable, List
 import argparse
+import logging as log
 import textwrap
 
 
@@ -69,3 +70,11 @@ def join_cli_args(args: argparse.Namespace, *names: Iterable[str]) -> List:
         joined += value
 
     return joined
+
+def show_video_import_warning():
+    log.warning("Using 'video_frames' in a project may lead "
+        "to different results across multiple runs, if the "
+        "system setup changes (library version, OS, etc.). "
+        "If you need stable results, consider splitting the video "
+        "manually using instructions at: "
+        "https://openvinotoolkit.github.io/datumaro/docs/user-manual/media_formats/")