+++ tests/test_video.py
@@ -1,12 +1,17 @@
+from unittest import TestCase
 import os.path as osp
 
 import cv2
 import numpy as np
 import pytest
 
+from datumaro.components.dataset import Dataset
+from datumaro.components.extractor import DatasetItem
 from datumaro.components.media import Video
-from datumaro.util.scope import on_exit_do, scoped
-from datumaro.util.test_utils import TestDir
+from datumaro.components.media_manager import MediaManager
+from datumaro.components.project import Project
+from datumaro.util.scope import Scope, on_exit_do, scope_add, scoped
+from datumaro.util.test_utils import TestDir, compare_datasets
 
 from .requirements import Requirements, mark_requirement
 
@@ -30,7 +35,7 @@ def make_sample_video(path, frames=4, frame_size=(10, 20), fps=25.0):
 def fxt_sample_video():
     with TestDir() as test_dir:
         video_path = osp.join(test_dir, 'video.avi')
-        make_sample_video(video_path, frame_size=(4, 6))
+        make_sample_video(video_path, frame_size=(4, 6), frames=4)
 
         yield video_path
 
@@ -120,3 +125,100 @@ class VideoTest:
 
             assert osp.join(test_dir, 'path.mp4') == video.path
             assert '.mp4' == video.ext
+
+
+class VideoExtractorTest:
+    @mark_requirement(Requirements.DATUM_GENERAL_REQ)
+    @scoped
+    def test_can_read_frames(self, fxt_sample_video):
+        on_exit_do(MediaManager.get_instance().clear)
+
+        expected = Dataset.from_iterable([
+            DatasetItem('frame_%03d' % i, subset='train',
+                image=np.ones((4, 6, 3)) * i)
+            for i in range(4)
+        ])
+
+        actual = Dataset.import_from(fxt_sample_video, 'video_frames',
+            subset='train', name_pattern='frame_%03d')
+
+        compare_datasets(TestCase(), expected, actual)
+
+    @mark_requirement(Requirements.DATUM_GENERAL_REQ)
+    @scoped
+    def test_can_split_and_load(self, fxt_sample_video):
+        test_dir = scope_add(TestDir())
+        on_exit_do(MediaManager.get_instance().clear)
+
+        expected = Dataset.from_iterable([
+            DatasetItem('frame_%06d' % i, image=np.ones((4, 6, 3)) * i)
+            for i in range(4)
+        ])
+
+        dataset = Dataset.import_from(fxt_sample_video, 'video_frames',
+            start_frame=0, end_frame=4, name_pattern='frame_%06d')
+        dataset.export(format='image_dir', save_dir=test_dir,
+            image_ext='.jpg')
+
+        actual = Dataset.import_from(test_dir, 'image_dir')
+        compare_datasets(TestCase(), expected, actual)
+
+class ProjectTest:
+    @mark_requirement(Requirements.DATUM_GENERAL_REQ)
+    def test_can_release_resources_on_exit(self, fxt_sample_video):
+        with Scope() as scope:
+            test_dir = scope.add(TestDir())
+
+            project = scope.add(Project.init(test_dir))
+
+            project.import_source('src', osp.dirname(fxt_sample_video),
+                'video_frames', rpath=osp.basename(fxt_sample_video))
+
+            assert len(project.working_tree.make_dataset()) == 4
+        assert not osp.exists(test_dir)
+
+    @mark_requirement(Requirements.DATUM_GENERAL_REQ)
+    @scoped
+    def test_can_release_resources_on_remove(self, fxt_sample_video):
+        test_dir = scope_add(TestDir())
+
+        project = scope_add(Project.init(test_dir))
+
+        project.import_source('src', osp.dirname(fxt_sample_video),
+            'video_frames', rpath=osp.basename(fxt_sample_video))
+        project.commit('commit 1')
+
+        assert len(project.working_tree.make_dataset()) == 4
+        assert osp.isdir(osp.join(test_dir, 'src'))
+
+        project.remove_source('src', keep_data=False)
+
+        assert not osp.exists(osp.join(test_dir, 'src'))
+
+    @mark_requirement(Requirements.DATUM_GENERAL_REQ)
+    @scoped
+    def test_can_release_resources_on_checkout(self, fxt_sample_video):
+        test_dir = scope_add(TestDir())
+
+        project = scope_add(Project.init(test_dir))
+
+        src_url = osp.join(test_dir, 'src')
+        src = Dataset.from_iterable([
+            DatasetItem(1),
+        ], categories=['a'])
+        src.save(src_url)
+        project.add_source(src_url, 'datumaro')
+        project.commit('commit 1')
+
+        project.remove_source('src', keep_data=False)
+
+        project.import_source('src', osp.dirname(fxt_sample_video),
+            'video_frames', rpath=osp.basename(fxt_sample_video))
+        project.commit('commit 2')
+
+        assert len(project.working_tree.make_dataset()) == 4
+        assert osp.isdir(osp.join(test_dir, 'src'))
+
+        project.checkout('HEAD~1')
+
+        assert len(project.working_tree.make_dataset()) == 1